+++ modules/core/src/cpp/Pause.cpp
@@ -20,7 +20,7 @@ Pause(Evaluator* eval, double seconds)
 {
     if (std::isinf(seconds)) {
         while (!NelsonConfiguration::getInstance()->getInterruptPending()) {
-            std::this_thread::sleep_for(std::chrono::milliseconds(uint64(10)));
+            std::this_thread::sleep_for(std::chrono::milliseconds(uint64(1)));
             if (eval && eval->haveEventsLoop()) {
                 ProcessEventsDynamicFunctionWithoutWait();
             }
@@ -32,7 +32,7 @@ Pause(Evaluator* eval, double seconds)
             = std::chrono::high_resolution_clock::now().time_since_epoch();
         bool bContinue = true;
         do {
-            std::this_thread::sleep_for(std::chrono::nanoseconds(uint64(10)));
+            std::this_thread::sleep_for(std::chrono::milliseconds(uint64(1)));
             std::chrono::nanoseconds current_time
                 = std::chrono::high_resolution_clock::now().time_since_epoch();
             std::chrono::nanoseconds difftime = (current_time - begin_time);